# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Hôtel API
  description: API de gestion hôtelière — SAÉ S6 IUT Rodez
  version: 1.0.0
  contact:
    name: Alexandre Brouzes
servers:
  - url: http://localhost:8060
    description: Serveur local
...
paths:
...
  /room-types:
    get:
      tags: [Room Types]
      summary: Lister tous les types de chambres
      operationId: getAllRoomTypes
      responses:
        '200':
          description: Liste des types de chambres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomTypeResponse'
              example:
                - id: 1
                  name: Chambre Simple
                  totalRooms: 10
                - id: 2
                  name: Chambre Double
                  totalRooms: 8
                - id: 3
                  name: Suite
                  totalRooms: 3

    post:
      tags: [Room Types]
      summary: Créer un nouveau type de chambre
      operationId: createRoomType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomTypeRequest'
            example:
              name: Suite Prestige
              totalRooms: 5
      responses:
        '201':
          description: Type de chambre créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomTypeResponse'
              example:
                id: 4
                name: Suite Prestige
                totalRooms: 5
        '400':
          $ref: '#/components/responses/BadRequest'
  /room-types/{id}:
    get:
      tags: [Room Types]
      summary: Récupérer un type de chambre par ID
      operationId: getRoomTypeById
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Type de chambre trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomTypeResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── US-02 : Tarifs ───────────────────────────────────────────────────────
  /room-types/{id}/prices:
    post:
      tags: [Prices]
      summary: Ajouter un tarif par période à un type de chambre
      operationId: addPriceToRoomType
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomTypePriceRequest'
            example:
              startDate: '2026-07-01'
              endDate: '2026-08-31'
              pricePerNight: 129.00
      responses:
        '201':
          description: Tarif ajouté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomTypePriceResponse'
              example:
                id: 8
                roomTypeId: 1
                startDate: '2026-07-01'
                endDate: '2026-08-31'
                pricePerNight: 129.00
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── US-03 : Disponibilité ────────────────────────────────────────────────
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      example: 1
  responses:
    BadRequest:
      description: Données invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    RoomTypeRequest:
      type: object
      required: [name, totalRooms]
      properties:
        name:
          type: string
          example: Suite Prestige
        totalRooms:
          type: integer
          minimum: 1
          example: 5
    RoomTypePriceRequest:
      type: object
      required: [startDate, endDate, pricePerNight]
      properties:
        startDate:
          type: string
          format: date
          example: '2026-07-01'
        endDate:
          type: string
          format: date
          example: '2026-08-31'
        pricePerNight:
          type: number
          format: double
          minimum: 0.01
          example: 129.00
...
    RoomTypeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        totalRooms:
          type: integer
    RoomTypePriceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        roomTypeId:
          type: integer
          format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        pricePerNight:
          type: number
          format: double
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time